Async parsing of packages in SBOM
