+++ packages/backend/src/megalinter/parsers/parse-sbom.ts
@@ -1,13 +1,11 @@
-import axios from "axios";
-import { SbomPackage, Severity } from "shared-types";
-import licenseConfig from "../../sbom/licenseConfig.json";
-import { ReportStore } from "../../stores/be-report-store";
+import { SbomPackage } from "shared-types";
+import { Component, Dependency } from "../../sbom/sbom-types";
 import {
-  Component,
-  Dependency,
-  LicenseInfo,
-  NpmLicense,
-} from "../megalinter-sbom-types";
+  extractMappingForApplication,
+  fetchPackages,
+  getPackages,
+} from "../../sbom/sbom-utils";
+import { ReportStore } from "../../stores/be-report-store";
 import { LinterCompleteMessage } from "../megalinter-types";
 
 export const parseSBOM = async (
@@ -28,10 +26,12 @@ export const parseSBOM = async (
       const dependencies = msg.outputSarif.runs[0].properties.megalinter.sbom
         .dependencies as Dependency[];
       const applications = extractMappingForApplication(components);
+      const pkgsInfo = await fetchPackages(dependencies, components);
       const sbomPackages: SbomPackage[] = await getPackages(
         dependencies,
         components,
-        applications
+        applications,
+        pkgsInfo
       );
       if (sbomPackages && sbomPackages.length > 0) {
         reportStore.set({ packages: sbomPackages });
@@ -39,155 +39,3 @@ export const parseSBOM = async (
     }
   }
 };
-
-function sortByLicenseLength(arr: LicenseInfo[]): LicenseInfo[] {
-  return arr.sort((a, b) => b.license.length - a.license.length);
-}
-
-async function getPackages(
-  dependencies: Dependency[],
-  components: Component[],
-  applications: Record<string, string>
-) {
-  const sbomPackages: SbomPackage[] = [];
-  const sortedLicenseConfig = sortByLicenseLength(licenseConfig);
-  let filePath = "";
-
-  for (const dependency of dependencies) {
-    // if (applications.hasOwnProperty(dependency.ref)) {
-    if (Object.prototype.hasOwnProperty.call(applications, dependency.ref)) {
-      filePath = applications[dependency.ref];
-    } else {
-      console.log(`no application for ref: ${dependency.ref}`);
-    }
-    if (dependency.dependsOn) {
-      for (const purl of dependency.dependsOn) {
-        console.log(`working on purl ${purl}`);
-        const component = components.find(
-          (component) => component.purl === purl
-        );
-        let registry = "";
-        let license = "Unknown";
-        let severity = Severity.Medium;
-        if (component) {
-          const packageName = component.name;
-          const packageVersion = component.version;
-          const sourceList: string[] = [];
-
-          if (purl.startsWith("pkg:pypi")) {
-            registry = "PyPi";
-            try {
-              const packageInfo = await fetchDataFromPyPi(
-                packageName,
-                packageVersion
-              );
-              if (packageInfo?.info?.license) {
-                sourceList.push(packageInfo.info.license);
-              }
-              if (packageInfo.info.classifiers) {
-                sourceList.push(packageInfo?.info.classifiers.join(" "));
-              }
-            } catch (error) {
-              console.error("Error:", error);
-            }
-          } else if (purl.startsWith("pkg:npm")) {
-            try {
-              const packageInfo = await fetchDataFromNPM(
-                packageName,
-                packageVersion
-              );
-              if (packageInfo?.license) {
-                sourceList.push(packageInfo.license);
-              } else if (packageInfo?.licenses) {
-                packageInfo.licenses.forEach((npmLicense: NpmLicense) => {
-                  if (npmLicense.type) {
-                    sourceList.push(npmLicense.type);
-                  }
-                });
-              } else {
-                console.log(`missing license for: ${purl}`);
-              }
-            } catch (error) {
-              console.error("Error:", error);
-            }
-          } else {
-            console.log(`purl: ${purl}`);
-          }
-
-          if (sourceList.length == 0) {
-            console.log(`no where to get license for ${packageName}`);
-          } else {
-            for (const licenseSoruce of sourceList) {
-              const licenseItem = sortedLicenseConfig.find((item) =>
-                licenseSoruce.includes(item.license)
-              );
-              if (licenseItem) {
-                license = licenseItem.license;
-                severity = Severity[licenseItem.severity];
-                break;
-              }
-            }
-          }
-
-          sbomPackages.push({
-            packageName: component.name,
-            packageVersion: component.version,
-            license: license,
-            registry: registry,
-            severity: severity,
-            filePath: filePath,
-          });
-        } else {
-          console.log(`missing component info, purl: ${purl}`);
-        }
-      }
-    } else {
-      console.log(`no depends for ref: ${dependency.ref}`);
-    }
-  }
-
-  const severityOrder = ["critical", "high", "medium", "low"];
-  sbomPackages.sort(
-    (a, b) =>
-      severityOrder.indexOf(b.severity) - severityOrder.indexOf(a.severity)
-  );
-  return sbomPackages;
-}
-
-// eslint-disable-next-line @typescript-eslint/no-explicit-any
-async function fetchDataFromPyPi(name: string, version: string): Promise<any> {
-  try {
-    const url = `https://pypi.org/pypi/${name}/${version}/json`;
-    const response = await axios.get(url);
-    return response.data;
-  } catch (error) {
-    console.error("Error fetching data from PyPi:", error);
-    throw error;
-  }
-}
-
-// eslint-disable-next-line @typescript-eslint/no-explicit-any
-async function fetchDataFromNPM(name: string, version: string): Promise<any> {
-  try {
-    const url = `https://registry.npmjs.org/${name}/${version}`;
-    const response = await axios.get(url);
-    return response.data;
-  } catch (error) {
-    console.error("Error fetching data from NPM:", error);
-    throw error;
-  }
-}
-
-function extractMappingForApplication(
-  components: Component[]
-): Record<string, string> {
-  const mapping: Record<string, string> = {};
-
-  for (const component of components) {
-    if (component.type === "application") {
-      mapping[component["bom-ref"]] = component.name;
-    }
-  }
-
-  return mapping;
-}
